refactor: split out ealloc module
